feat(ProductClient): encrypt user information
